+++ src/main/java/org/ethereumclassic/etherjar/rpc/json/TransactionLogJsonDeserializer.java
@@ -26,16 +26,16 @@ public class TransactionLogJsonDeserializer extends EtherJsonDeserializer<Transa
 
         log.setAddress(getAddress(node, "address"));
         log.setBlockHash(getBlockHash(node, "blockHash"));
-        log.setBlockNumber(getQuantity(node, "blockNumber").getValue().intValue());
+        log.setBlockNumber(getQuantity(node, "blockNumber").getValue().longValue());
         log.setData(getData(node, "data"));
-        log.setLogIndex(getQuantity(node, "logIndex").getValue().intValue());
+        log.setLogIndex(getQuantity(node, "logIndex").getValue().longValue());
         List<HexData> topics = new ArrayList<>();
         for (JsonNode topic: node.get("topics")) {
             topics.add(HexData.from(topic.textValue()));
         }
         log.setTopics(topics);
         log.setTransactionHash(getTxHash(node, "transactionHash"));
-        log.setTransactionIndex(getQuantity(node, "transactionIndex").getValue().intValue());
+        log.setTransactionIndex(getQuantity(node, "transactionIndex").getValue().longValue());
 
         return log;
     }
